feat: 펀딩/위시풀 상태 선택 UI (#18)
